feat: Enhance testing framework by adding UI tests and updating test targets; refactor conversation model for improved message handling
