+++ .github/workflows/run-tests.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        test_target: [unit-test, intg-test, build-test]
+        test_target: [unit-test, ui-test, intg-test, build-test]
       fail-fast: false
 
     steps:
