+++ pyproject.toml
@@ -8,7 +8,7 @@ authors = [
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "olm-api @ git+https://github.com/akitorahayashi/olm-api.git@v2.2.0",
+    "olm-api @ git+https://github.com/akitorahayashi/olm-api.git@v2.2.2",
     "ollama>=0.4.4",
     "streamlit>=1.49.1",
     "pandas>=2.0.0",
