--- src/schemas/__init__.py
@@ -1,20 +1,0 @@
-from .message import Message, MessageRole
-from .response import (
-    ChatResponse,
-    ChatStreamResponse,
-    Choice,
-    StreamChoice,
-    StreamDelta,
-    Usage,
-)
-
-__all__ = [
-    "Message",
-    "MessageRole",
-    "ChatResponse",
-    "ChatStreamResponse",
-    "Choice",
-    "StreamChoice",
-    "StreamDelta",
-    "Usage",
-]
