+++ tests/unit/models/test_conversation_model.py
@@ -3,8 +3,8 @@ from unittest.mock import AsyncMock, MagicMock, mock_open, patch
 
 import pytest
 
+from olm_api.api.v2.schemas.message import Message, MessageRole
 from src.models.conversation_model import ConversationModel
-from src.schemas import Message, MessageRole
 
 
 @pytest.fixture
@@ -37,7 +37,7 @@ class TestConversationModel:
     def test_add_ai_message(self, conversation_model):
         """Test that an AI message is added correctly."""
         conversation_model.add_ai_message("Hi there")
-        assert conversation_model.messages == [{"role": "ai", "content": "Hi there"}]
+        assert conversation_model.messages == [{"role": "assistant", "content": "Hi there"}]
 
     def test_reset(self, conversation_model):
         """Test that the reset method clears messages."""
